+++ src/main/java/io/codiga/plugins/jetbrains/completion/CodigaCompletion.java
@@ -15,3 +15,6 @@ public class CodigaCompletion extends CompletionContributor {
             new CodigaCompletionProvider());
     }
 }
+
+
+